Add AddCoAuthor method to commit template
The case where a preexisting co-author declaration is not the last
line of the template is not supported yet. It would work, but the
co-auhtor declarations wouldn't be grouped and the new one would
be added immediately after the last line of the template.

Potentially, it could even be appended to the last line of the template
if the template didn't previously end with a new line.