Refactor reorganize project as library and command
I prefer keeping the CLI domain scoped to user interaction because
it is more difficult to automatically test.
With the core functionality provided as a library, most of the application
behaviour gets to be tested.

I wasn't sure how to do it in Go, but this pattern makes sense to me : )
See https://medium.com/@benbjohnson/structuring-applications-in-go-3b04be4ff091